Delete report-example-2.pdf
No longer meets the project requirements.

Based on the feedback ""To begin with, the report does not follow the template. The sections are in jumbled up, and the titles of many sections do not match the template either. Then there are some sections missing like the benchmark, conclusion (final conclusion is missing), justification, etc. Some sections do not have enough detail (eg: the choice of metric is not justified). I did not go through the entire report in depth, so there might be some other issues that I have missed as well. I do not think this is a very good template report to be shared with the students. It is well written, but perhaps not with the required detail and format that the reviewers are expecting for this project.""